requirements.txt
# ── Core ───────────────────────────────────────────────────────────────────
pyyaml==6.0.2
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.2
aiofiles==23.2.1

# ── Voice Pipeline ─────────────────────────────────────────────────────────
faster-whisper==1.1.0
sounddevice==0.5.1
soundfile==0.12.1
numpy==1.26.4
openwakeword==0.6.0

# ── LLM / Brain ────────────────────────────────────────────────────────────
ollama==0.3.3

# ── Vision ─────────────────────────────────────────────────────────────────
ultralytics==8.3.0
mediapipe==0.10.14
opencv-python==4.10.0.84
Pillow==10.4.0
aiohttp==3.13.5

# ── Audio Classification ───────────────────────────────────────────────────
tensorflow==2.17.0
tensorflow-hub==0.16.1

# ── MQTT ───────────────────────────────────────────────────────────────────
paho-mqtt==2.1.0
aiomqtt==2.5.1

# ── Database ───────────────────────────────────────────────────────────────
aiosqlite==0.20.0

# ── Windows Process Monitor ────────────────────────────────────────────────
pywin32==306
psutil==6.1.0

# ── Dashboard (real-time web UI) ───────────────────────────────────────────
fastapi==0.115.0
uvicorn==0.32.0
websockets==13.1

# ── Testing ────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
